requirements.txt
flask
mysql-connector-python
psycopg2-binary
dict2xml